refactor: split forwarding build from runtime resolution
Move snapshot compilation functions into forwarding_build.rs (605 lines).
forwarding.rs drops from 3806 to 3199 lines. Pure refactor.

Moved: build_forwarding_state, build_screen_profiles, pick_interface_*,
resolve_route_target_*, parse_route_next_hop*, resolve_ifindex,
infer_connected_route_target_*.

Closes #368
