+++ userspace-dp/src/afxdp/forwarding.rs
@@ -1,5 +1,4 @@
 use super::*;
-use crate::RouteSnapshot;
 
 const DEFAULT_V4_TABLE: &str = "inet.0";
 const DEFAULT_V6_TABLE: &str = "inet6.0";
@@ -24,434 +23,6 @@ pub(super) fn classify_metadata(
     }
 }
 
-pub(super) fn build_screen_profiles(snapshot: &ConfigSnapshot) -> FxHashMap<String, ScreenProfile> {
-    let mut profiles = FxHashMap::default();
-    for sp in &snapshot.screens {
-        if sp.zone.is_empty() {
-            continue;
-        }
-        profiles.insert(
-            sp.zone.clone(),
-            ScreenProfile {
-                land: sp.land,
-                syn_fin: sp.syn_fin,
-                no_flag: sp.tcp_no_flag,
-                fin_no_ack: sp.fin_no_ack,
-                winnuke: sp.winnuke,
-                ping_death: sp.ping_death,
-                teardrop: sp.teardrop,
-                icmp_fragment: sp.icmp_fragment,
-                source_route: sp.source_route,
-                icmp_flood_threshold: sp.icmp_flood_threshold,
-                udp_flood_threshold: sp.udp_flood_threshold,
-                syn_flood_threshold: sp.syn_flood_threshold,
-                session_limit_src: sp.session_limit_src,
-                session_limit_dst: sp.session_limit_dst,
-                port_scan_threshold: sp.port_scan_threshold,
-                ip_sweep_threshold: sp.ip_sweep_threshold,
-            },
-        );
-    }
-    profiles
-}
-
-pub(super) fn build_forwarding_state(snapshot: &ConfigSnapshot) -> ForwardingState {
-    let mut state = ForwardingState::default();
-    let mut name_to_ifindex = BTreeMap::new();
-    let mut linux_to_ifindex = BTreeMap::new();
-    let mut mac_by_ifindex = BTreeMap::new();
-    let (excluded_local_v4, excluded_local_v6) = nat_translated_local_exclusions(snapshot);
-
-    for zone in &snapshot.zones {
-        if zone.id == 0 || zone.name.is_empty() {
-            continue;
-        }
-        state.zone_name_to_id.insert(zone.name.clone(), zone.id);
-        state.zone_id_to_name.insert(zone.id, zone.name.clone());
-    }
-
-    for endpoint in &snapshot.tunnel_endpoints {
-        if endpoint.id == 0 || endpoint.ifindex <= 0 {
-            continue;
-        }
-        let Ok(source) = endpoint.source.parse::<IpAddr>() else {
-            continue;
-        };
-        let Ok(destination) = endpoint.destination.parse::<IpAddr>() else {
-            continue;
-        };
-        let outer_family = match (endpoint.outer_family.as_str(), destination) {
-            ("inet6", _) => libc::AF_INET6,
-            ("inet", _) => libc::AF_INET,
-            (_, IpAddr::V6(_)) => libc::AF_INET6,
-            _ => libc::AF_INET,
-        };
-        let transport_table =
-            canonical_route_table(&endpoint.transport_table, outer_family == libc::AF_INET6);
-        state.tunnel_endpoints.insert(
-            endpoint.id,
-            TunnelEndpoint {
-                id: endpoint.id,
-                logical_ifindex: endpoint.ifindex,
-                redundancy_group: endpoint.redundancy_group,
-                mode: endpoint.mode.clone(),
-                outer_family,
-                source,
-                destination,
-                key: endpoint.key,
-                ttl: endpoint.ttl.max(0) as u8,
-                transport_table,
-            },
-        );
-        state
-            .tunnel_endpoint_by_ifindex
-            .insert(endpoint.ifindex, endpoint.id);
-    }
-
-    for iface in &snapshot.interfaces {
-        if iface.ifindex <= 0 {
-            continue;
-        }
-        let label = if iface.linux_name.is_empty() {
-            iface.name.clone()
-        } else {
-            iface.linux_name.clone()
-        };
-        state.ifindex_to_name.insert(iface.ifindex, label);
-        name_to_ifindex.insert(iface.name.clone(), iface.ifindex);
-        if !iface.linux_name.is_empty() {
-            linux_to_ifindex.insert(iface.linux_name.clone(), iface.ifindex);
-        }
-        if !iface.zone.is_empty() {
-            state
-                .ifindex_to_zone
-                .insert(iface.ifindex, iface.zone.clone());
-            if iface.parent_ifindex > 0 {
-                match state.ifindex_to_zone.get(&iface.parent_ifindex) {
-                    Some(existing) if existing != &iface.zone => {}
-                    _ => {
-                        state
-                            .ifindex_to_zone
-                            .insert(iface.parent_ifindex, iface.zone.clone());
-                    }
-                }
-            }
-        }
-        if iface.tunnel {
-            state.tunnel_interfaces.insert(iface.ifindex);
-        }
-        if let Some(mac) = parse_mac(&iface.hardware_addr) {
-            mac_by_ifindex.insert(iface.ifindex, mac);
-        }
-        let tunnel_endpoint_id = state
-            .tunnel_endpoint_by_ifindex
-            .get(&iface.ifindex)
-            .copied()
-            .unwrap_or(0);
-        for addr in &iface.addresses {
-            let Ok(net) = addr.address.parse::<IpNet>() else {
-                continue;
-            };
-            match net {
-                IpNet::V4(v4) => {
-                    if excluded_local_v4.contains(&v4.addr()) {
-                        state.interface_nat_v4.insert(v4.addr(), iface.ifindex);
-                    } else {
-                        state.local_v4.insert(v4.addr());
-                    }
-                    state.connected_v4.push(ConnectedRouteV4 {
-                        prefix: PrefixV4::from_net(v4),
-                        ifindex: iface.ifindex,
-                        tunnel_endpoint_id,
-                    });
-                }
-                IpNet::V6(v6) => {
-                    if excluded_local_v6.contains(&v6.addr()) {
-                        state.interface_nat_v6.insert(v6.addr(), iface.ifindex);
-                    } else {
-                        state.local_v6.insert(v6.addr());
-                    }
-                    state.connected_v6.push(ConnectedRouteV6 {
-                        prefix: PrefixV6::from_net(v6),
-                        ifindex: iface.ifindex,
-                        tunnel_endpoint_id,
-                    });
-                }
-            }
-        }
-    }
-
-    for iface in &snapshot.interfaces {
-        if iface.ifindex <= 0 {
-            continue;
-        }
-        let bind_ifindex = if iface.parent_ifindex > 0 {
-            iface.parent_ifindex
-        } else {
-            iface.ifindex
-        };
-        let src_mac = match parse_mac(&iface.hardware_addr)
-            .or_else(|| mac_by_ifindex.get(&bind_ifindex).copied())
-            .or_else(|| iface.tunnel.then_some([0; 6]))
-        {
-            Some(mac) => mac,
-            None => continue,
-        };
-        state.egress.insert(
-            iface.ifindex,
-            EgressInterface {
-                bind_ifindex,
-                vlan_id: iface.vlan_id.max(0) as u16,
-                mtu: iface.mtu.max(0) as usize,
-                src_mac,
-                zone: iface.zone.clone(),
-                redundancy_group: iface.redundancy_group,
-                primary_v4: pick_interface_v4(iface),
-                primary_v6: pick_interface_v6(iface),
-            },
-        );
-    }
-
-    state
-        .connected_v4
-        .sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
-    state
-        .connected_v6
-        .sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
-
-    for route in &snapshot.routes {
-        if let Ok(prefix) = route.destination.parse::<Ipv4Net>() {
-            let (next_hop, ifindex, tunnel_endpoint_id) =
-                resolve_route_target_v4(route, &name_to_ifindex, &linux_to_ifindex, &state);
-            let table = canonical_route_table(&route.table, false);
-            state
-                .routes_v4
-                .entry(table)
-                .or_default()
-                .push(RouteEntryV4 {
-                    prefix: PrefixV4::from_net(prefix),
-                    ifindex,
-                    tunnel_endpoint_id,
-                    next_hop,
-                    discard: route.discard,
-                    next_table: route.next_table.clone(),
-                });
-            continue;
-        }
-        if let Ok(prefix) = route.destination.parse::<Ipv6Net>() {
-            let (next_hop, ifindex, tunnel_endpoint_id) =
-                resolve_route_target_v6(route, &name_to_ifindex, &linux_to_ifindex, &state);
-            let table = canonical_route_table(&route.table, true);
-            state
-                .routes_v6
-                .entry(table)
-                .or_default()
-                .push(RouteEntryV6 {
-                    prefix: PrefixV6::from_net(prefix),
-                    ifindex,
-                    tunnel_endpoint_id,
-                    next_hop,
-                    discard: route.discard,
-                    next_table: route.next_table.clone(),
-                });
-        }
-    }
-    for routes in state.routes_v4.values_mut() {
-        routes.sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
-    }
-    for routes in state.routes_v6.values_mut() {
-        routes.sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
-    }
-
-    for neigh in &snapshot.neighbors {
-        if neigh.ifindex <= 0 || !neighbor_state_usable(&neigh.state) {
-            continue;
-        }
-        let Ok(ip) = neigh.ip.parse::<IpAddr>() else {
-            continue;
-        };
-        let Some(mac) = parse_mac(&neigh.mac) else {
-            continue;
-        };
-        state
-            .neighbors
-            .insert((neigh.ifindex, ip), NeighborEntry { mac });
-    }
-    for fabric in &snapshot.fabrics {
-        if fabric.parent_ifindex <= 0 {
-            continue;
-        }
-        let Ok(peer_addr) = fabric.peer_address.parse::<IpAddr>() else {
-            continue;
-        };
-        let local_mac = parse_mac(&fabric.local_mac)
-            .or_else(|| mac_by_ifindex.get(&fabric.parent_ifindex).copied());
-        let Some(local_mac) = local_mac else {
-            continue;
-        };
-        let peer_mac = parse_mac(&fabric.peer_mac).or_else(|| {
-            state
-                .neighbors
-                .get(&(fabric.overlay_ifindex, peer_addr))
-                .or_else(|| state.neighbors.get(&(fabric.parent_ifindex, peer_addr)))
-                .map(|entry| entry.mac)
-        });
-        let Some(peer_mac) = peer_mac else {
-            continue;
-        };
-        state.fabrics.push(FabricLink {
-            parent_ifindex: fabric.parent_ifindex,
-            overlay_ifindex: fabric.overlay_ifindex,
-            peer_addr,
-            peer_mac,
-            local_mac,
-        });
-    }
-    state.policy = parse_policy_state(&snapshot.default_policy, &snapshot.policies);
-    state.allow_dns_reply = snapshot.flow.allow_dns_reply;
-    state.allow_embedded_icmp = snapshot.flow.allow_embedded_icmp;
-    state.session_timeouts = crate::session::SessionTimeouts::from_seconds(
-        snapshot.flow.tcp_session_timeout,
-        snapshot.flow.udp_session_timeout,
-        snapshot.flow.icmp_session_timeout,
-    );
-    state.source_nat_rules = parse_source_nat_rules(&snapshot.source_nat_rules);
-    state.static_nat = StaticNatTable::from_snapshots(&snapshot.static_nat_rules);
-    state.dnat_table = DnatTable::from_snapshots(&snapshot.destination_nat_rules);
-    state.nat64 = Nat64State::from_snapshots(&snapshot.nat64_rules);
-    state.nptv6 = Nptv6State::from_snapshots(&snapshot.nptv6_rules);
-    state.screen_profiles = build_screen_profiles(snapshot);
-    state.tcp_mss_all_tcp = snapshot.flow.tcp_mss_all_tcp;
-    state.tcp_mss_ipsec_vpn = snapshot.flow.tcp_mss_ipsec_vpn;
-    state.tcp_mss_gre_in = snapshot.flow.tcp_mss_gre_in;
-    state.tcp_mss_gre_out = snapshot.flow.tcp_mss_gre_out;
-    // Build filter state from snapshot
-    state.filter_state = crate::filter::parse_filter_state(
-        &snapshot.filters,
-        &snapshot.policers,
-        &snapshot.interfaces,
-        &snapshot.flow.lo0_filter_input_v4,
-        &snapshot.flow.lo0_filter_input_v6,
-    );
-    // Build flow export config from snapshot
-    state.flow_export_config = snapshot.flow_export.as_ref().and_then(|fe| {
-        let addr = format!("{}:{}", fe.collector_address, fe.collector_port);
-        addr.parse::<std::net::SocketAddr>().ok().map(|collector| {
-            crate::flowexport::FlowExportConfig {
-                collector,
-                sampling_rate: fe.sampling_rate,
-                active_timeout_secs: fe.active_timeout as u64,
-                inactive_timeout_secs: fe.inactive_timeout as u64,
-            }
-        })
-    });
-
-    // Add static NAT external IPs as local delivery targets so inbound
-    // traffic destined to external IPs is recognized by the firewall.
-    for ext_ip in state.static_nat.external_ips() {
-        match ext_ip {
-            IpAddr::V4(v4) => {
-                state.local_v4.insert(*v4);
-            }
-            IpAddr::V6(v6) => {
-                state.local_v6.insert(*v6);
-            }
-        }
-    }
-
-    // Add DNAT destination IPs as local delivery targets so traffic
-    // to those IPs is recognized as locally-destined and processed.
-    for dst_ip in state.dnat_table.destination_ips() {
-        match dst_ip {
-            IpAddr::V4(v4) => {
-                state.local_v4.insert(v4);
-            }
-            IpAddr::V6(v6) => {
-                state.local_v6.insert(v6);
-            }
-        }
-    }
-
-    // Debug: dump zone mappings and policy rules
-    #[cfg(feature = "debug-log")]
-    {
-        debug_log!("FWD_STATE: ifindex_to_zone={:?}", state.ifindex_to_zone);
-        debug_log!(
-            "FWD_STATE: egress keys={:?}",
-            state.egress.keys().collect::<Vec<_>>()
-        );
-        for (ifidx, eg) in &state.egress {
-            debug_log!(
-                "FWD_STATE: egress[{}] bind={} zone={} vlan={} mtu={}",
-                ifidx,
-                eg.bind_ifindex,
-                eg.zone,
-                eg.vlan_id,
-                eg.mtu,
-            );
-        }
-        debug_log!(
-            "FWD_STATE: policy default={:?} rules={}",
-            state.policy.default_action,
-            state.policy.rules.len(),
-        );
-        for (i, rule) in state.policy.rules.iter().enumerate() {
-            debug_log!(
-                "FWD_STATE: policy[{}] {}->{}  action={:?} src_v4={} dst_v4={} apps={}",
-                i,
-                rule.from_zone,
-                rule.to_zone,
-                rule.action,
-                rule.source_v4.len(),
-                rule.destination_v4.len(),
-                rule.applications.len(),
-            );
-        }
-        debug_log!(
-            "FWD_STATE: local_v4={:?} interface_nat_v4={:?}",
-            state.local_v4,
-            state.interface_nat_v4,
-        );
-        debug_log!(
-            "FWD_STATE: snat_rules={} static_nat={} dnat_table={} nptv6={} connected_v4={} routes_v4={}",
-            state.source_nat_rules.len(),
-            if state.static_nat.is_empty() {
-                0
-            } else {
-                state.static_nat.external_ips().count()
-            },
-            if state.dnat_table.is_empty() {
-                0
-            } else {
-                state.dnat_table.destination_ips().count()
-            },
-            if state.nptv6.is_empty() {
-                0
-            } else {
-                state.nptv6.external_prefixes().len()
-            },
-            state.connected_v4.len(),
-            state.routes_v4.values().map(|v| v.len()).sum::<usize>(),
-        );
-    }
-
-    // Install nftables rules to suppress kernel TCP RSTs from SNAT IPs.
-    //
-    // When the AF_XDP fill ring momentarily runs dry under high load,
-    // the mlx5 driver falls back to the regular RX path. Those leaked
-    // packets reach the kernel TCP stack which — having no matching
-    // socket — sends RSTs to the server, killing the connection.
-    // Blocking outgoing RSTs for SNAT-managed IPs is a targeted fix:
-    // the DP handles all TCP state for those addresses.
-    install_kernel_rst_suppression(&state);
-
-    state
-}
-
-/// Install nftables rules to DROP outgoing TCP RSTs from interface-NAT
-/// (SNAT) addresses.  These addresses are owned by the userspace
-/// dataplane; the kernel has no sockets for them and should never emit
-/// RSTs.
 pub(super) fn canonical_route_table(table: &str, is_ipv6: bool) -> String {
     if is_ipv6 {
         if table == "inet.0" {
@@ -469,185 +40,6 @@ pub(super) fn canonical_route_table(table: &str, is_ipv6: bool) -> String {
         return format!("{prefix}.inet.0");
     }
     table.to_string()
-}
-
-pub(super) fn pick_interface_v4(iface: &InterfaceSnapshot) -> Option<Ipv4Addr> {
-    let mut fallback = None;
-    for addr in &iface.addresses {
-        if addr.family != "inet" {
-            continue;
-        }
-        let Ok(net) = addr.address.parse::<Ipv4Net>() else {
-            continue;
-        };
-        let ip = net.addr();
-        if fallback.is_none() {
-            fallback = Some(ip);
-        }
-        if !ip.is_link_local() {
-            return Some(ip);
-        }
-    }
-    fallback
-}
-
-pub(super) fn pick_interface_v6(iface: &InterfaceSnapshot) -> Option<Ipv6Addr> {
-    let mut fallback = None;
-    for addr in &iface.addresses {
-        if addr.family != "inet6" {
-            continue;
-        }
-        let Ok(net) = addr.address.parse::<Ipv6Net>() else {
-            continue;
-        };
-        let ip = net.addr();
-        if fallback.is_none() {
-            fallback = Some(ip);
-        }
-        if !ip.is_unicast_link_local() {
-            return Some(ip);
-        }
-    }
-    fallback
-}
-
-pub(super) fn resolve_route_target_v4(
-    route: &RouteSnapshot,
-    names: &BTreeMap<String, i32>,
-    linux_names: &BTreeMap<String, i32>,
-    state: &ForwardingState,
-) -> (Option<Ipv4Addr>, i32, u16) {
-    if route.discard || !route.next_table.is_empty() {
-        return (None, 0, 0);
-    }
-    let Some((next_hop, interface)) = route
-        .next_hops
-        .first()
-        .map(|nh| parse_route_next_hop(nh.as_str()))
-    else {
-        return (None, 0, 0);
-    };
-    let target = interface
-        .as_deref()
-        .and_then(|name| resolve_ifindex(name, names, linux_names))
-        .map(|ifindex| {
-            (
-                ifindex,
-                state
-                    .tunnel_endpoint_by_ifindex
-                    .get(&ifindex)
-                    .copied()
-                    .unwrap_or(0),
-            )
-        })
-        .or_else(|| next_hop.and_then(|ip| infer_connected_route_target_v4(state, ip)));
-    let (ifindex, tunnel_endpoint_id) = target.unwrap_or((0, 0));
-    (next_hop, ifindex, tunnel_endpoint_id)
-}
-
-pub(super) fn resolve_route_target_v6(
-    route: &RouteSnapshot,
-    names: &BTreeMap<String, i32>,
-    linux_names: &BTreeMap<String, i32>,
-    state: &ForwardingState,
-) -> (Option<Ipv6Addr>, i32, u16) {
-    if route.discard || !route.next_table.is_empty() {
-        return (None, 0, 0);
-    }
-    let Some((next_hop, interface)) = route
-        .next_hops
-        .first()
-        .map(|nh| parse_route_next_hop_v6(nh.as_str()))
-    else {
-        return (None, 0, 0);
-    };
-    let target = interface
-        .as_deref()
-        .and_then(|name| resolve_ifindex(name, names, linux_names))
-        .map(|ifindex| {
-            (
-                ifindex,
-                state
-                    .tunnel_endpoint_by_ifindex
-                    .get(&ifindex)
-                    .copied()
-                    .unwrap_or(0),
-            )
-        })
-        .or_else(|| next_hop.and_then(|ip| infer_connected_route_target_v6(state, ip)));
-    let (ifindex, tunnel_endpoint_id) = target.unwrap_or((0, 0));
-    (next_hop, ifindex, tunnel_endpoint_id)
-}
-
-pub(super) fn parse_route_next_hop(spec: &str) -> (Option<Ipv4Addr>, Option<String>) {
-    let (ip_part, if_part) = if let Some((lhs, rhs)) = spec.split_once('@') {
-        (lhs, rhs)
-    } else {
-        (spec, "")
-    };
-    let ip = if ip_part.is_empty() {
-        None
-    } else {
-        ip_part.parse::<Ipv4Addr>().ok()
-    };
-    let iface = if if_part.is_empty() {
-        None
-    } else {
-        Some(if_part.to_string())
-    };
-    (ip, iface)
-}
-
-pub(super) fn parse_route_next_hop_v6(spec: &str) -> (Option<Ipv6Addr>, Option<String>) {
-    let (ip_part, if_part) = if let Some((lhs, rhs)) = spec.split_once('@') {
-        (lhs, rhs)
-    } else {
-        (spec, "")
-    };
-    let ip = if ip_part.is_empty() {
-        None
-    } else {
-        ip_part.parse::<Ipv6Addr>().ok()
-    };
-    let iface = if if_part.is_empty() {
-        None
-    } else {
-        Some(if_part.to_string())
-    };
-    (ip, iface)
-}
-
-pub(super) fn resolve_ifindex(
-    name: &str,
-    names: &BTreeMap<String, i32>,
-    linux_names: &BTreeMap<String, i32>,
-) -> Option<i32> {
-    names
-        .get(name)
-        .copied()
-        .or_else(|| linux_names.get(name).copied())
-}
-
-pub(super) fn infer_connected_route_target_v4(
-    state: &ForwardingState,
-    ip: Ipv4Addr,
-) -> Option<(i32, u16)> {
-    state
-        .connected_v4
-        .iter()
-        .find(|entry| entry.prefix.contains(ip))
-        .map(|entry| (entry.ifindex, entry.tunnel_endpoint_id))
-}
-
-pub(super) fn infer_connected_route_target_v6(
-    state: &ForwardingState,
-    ip: Ipv6Addr,
-) -> Option<(i32, u16)> {
-    state
-        .connected_v6
-        .iter()
-        .find(|entry| entry.prefix.contains(ip))
-        .map(|entry| (entry.ifindex, entry.tunnel_endpoint_id))
 }
 
 pub(super) fn neighbor_state_usable(state: &str) -> bool {
@@ -2147,6 +1539,7 @@ fn choose_v6_route(
 
 #[cfg(test)]
 mod tests {
+    use super::super::forwarding_build::*;
     use super::super::test_fixtures::*;
     use super::*;
     use crate::{FabricSnapshot, SourceNATRuleSnapshot};
